DevMode 2.5.1
Updated DevMode to latest version
