+++ Player/Blank.htm
@@ -7,9 +7,7 @@
     <script type="text/javascript" src="res://local/playercore.js"></script>
     <script type="text/javascript" src="res://local/desktopplayer.js"></script>
     <script type="text/javascript" src="res://local/paper.js"></script>
-    <script type="text/javascript" src="res://local/jquery.sidebar.min.js"></script>
-    <script type="text/javascript" src="res://local/tabulator.min.js"></script>
-    <script type="text/javascript" src="res://local/devmode_sidebar.js"></script>
+    <script type="text/javascript" src="res://local/DevMode.js"></script>
     <script type="text/paperscript" canvas="gridCanvas">
         // GRID_JS_PLACEHOLDER
     </script>
