+++ WorldModel/WorldModel/Core/CoreDevMode.aslx
@@ -6,7 +6,7 @@
   ****************************************************************************************************
   CoreDevMode - Tool for better testing the game
   by ScriptGames, 2018 - 2019
-  Version 2.5.0
+  Version 2.5.1
   ****************************************************************************************************
   Compatible with Quest 5.7 and higher
   ****************************************************************************************************
@@ -21,7 +21,11 @@
   Tool for better testing the game.
   Inspired by Pixie's Cheating-Code. ;-)
   ****************************************************************************************************
-  
+
+  ****************************************************************************************************
+  Version 2.5.1
+  Restructuring of the code.
+  Only the DevMode.aslx and DevMode.js are needed.
   ****************************************************************************************************
   Version 2.5.0
   Improved display of standbye mode
@@ -148,7 +152,6 @@
   ****************************************************************************************************
   -->
   <command name="devmode_commands">
-    <pattern type="string"></pattern>
     <pattern type="string"><![CDATA[^#(?<text>.*)]]></pattern>
     <script>
     <![CDATA[
